requirements.txt
kubernetes
PyYAML
jinja2

